slam_performance_model/requirements.txt
numpy==2.0.0
pandas==2.2.2
pillow==10.3.0
torch==2.3.1
torchvision==0.18.1
wandb==0.17.3
requests==2.32.3
python-dateutil==2.9.0.post0
six==1.16.0
typing_extensions==4.12.2
